add a hie.yaml to avoid the related lsp warning
ghcide etc. now use implicit-hie which in turn check for a hie.yaml file
in the current directory and if not present, resorts to using a default
hie.yaml file, but it doesn't actually create one, and continues with a
warning. The LSP i use (coc.nvim currently) seems to enjoy displaying a
long orange warning about this every time I open a file in a directory
without a hie.yaml, so while I can generate one with the `implicit-hie`
installed and the `gen-hie > ./hie.yaml` command I want to just have
something basic in thewre that will avoid this warning; can always just
edit it later to suit my then preferences.